:recycle: extract build logic to task
